pyproject.toml
[tool.poetry]
name = "trip-ingestor"
version = "1.0.0"
description = "Trip data ingestion service"
authors = ["Luiz Eduardo Malanczyn de Oliveira <oliveiraluiz.com@hotmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.16.0"
pandas = "^2.3.1"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
